webview_gui/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pywebview==5.1
